requirements.txt
# Core Dependencies
PyQt5==5.15.9
PyOpenGL==3.1.7
PyOpenGL-accelerate==3.1.7
numpy==1.24.3
scipy==1.11.4
matplotlib==3.8.2
pandas==2.1.4
pyqtgraph==0.13.3

# Optional but recommended
pillow==10.1.0
numba==0.58.1